+++ app/policies/product_policy.rb
@@ -1,8 +1,8 @@
 class ProductPolicy < ApplicationPolicy
   def show?    = barista? || any_manager? || prep_kitchen_manager? || prep_kitchen_worker?
   def index?   = show?
-  def create?  = privileged_manager?
-  def update?  = privileged_manager?
+  def create?  = uk_global_admin?
+  def update?  = uk_global_admin?
   def destroy? = uk_global_admin?
 
   class Scope < ApplicationPolicy::Scope